requirements.txt
kafka-python==2.2.8
selenium==4.27.1
PyYAML==6.0.1